mbnames: Allow custom sorting of mailboxes
mutt-sidebar shows mailboxes in the order they are listed in the file
written by mbnames. Therefore, to allow customization of the order with
which mailboxes are listed, introduce the new 'sortkey' directive in
the [mbnames] section.

'sortkey' specifies a function that will be called once for each
mailbox (passing a dict with 2 items: 'accountname' and 'foldername'),
and should return a suitable sort key that mailbox.

If 'sortkey' is not given, we default to the current alphabetical
sorting provided by using (d['accountname'], d['foldername']) as sort
key.

Signed-off-by: Johan Herland <johan@herland.net>