Add basic instructions on how to contribute to OfflineIMAP
I discovered in the pull request #95 that the development branch is
`next`, not `master`.

To prevent this kind of errors, GitHub shows a useful bright yellow banner
if a CONTRIBUTING.md file is present in the repository.
See <https://github.com/blog/1184-contributing-guidelines>.

I based this skeleton file on the instructions found at
<http://offlineimap.org/development.html>.