removed imaplib2 from this repo.
As the current version of imaplib2.py have an import restriction on the
names used in this repo, we have to copy and paste it in for the time
being. I've filed a pull request to remove that barrier, and I hope it
would be accepted.

It is to be noted that two features that we've introduced in offlineimap
is lost for the time being:

    1. traceback printing.
    2. repr used around various messages.

So this patch should be carefully evaluated before it can be
incorporated into offlineimap's codebase.