Use mail 'date' as file prefix if using utime_from_header.
The goal here is to use a timestamp based on mail Date (or Delivery-date
if not found) as filename prefix instead of current system time. This
allows to sort box by most recent mail based on file name. This is
useful in case of tools sorting mails using file name instead of mail
headers.