Use system encoding to decode passwords and usernames (fix #59)
u'\x00'.join() will implicitly decode all bytestrings that are provided,
using the default codec. The default codec being "ascii", this change
uses stdout encoding to try to decode the password and username. This
cannot provide worse compatibility than the existing behavior (which
breaks on any non-ascii character), and can be tweaked using the
environment in the edge cases.

Signed-off-by: mathieui <mathieui@mathieui.net>