code of conduct: try to clarify what item 3 might mean
I don't think the phrase "Free speech owns to anyone of us" is
grammatically correct, and I honestly don't know what the author
really meant, what does "owns" mean in this context? It could be that:

 - You should own up to what you've said

 - That we all "own" the right to free speech

I think the latter is probably more likely, in any case, submit a pull
request for this so this can be clarified.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Nicolas Sebrecht <nicolas.s-dev@laposte.net>