repository: Base: make nametrans consistency check stricter
We need to make sure that both transformations of folder names,
local -> remote -> local and remote -> local -> remote stay the same.

Commit 481efa95f6 started making use of the local to remote nametrans,
but only checked consistency for local -> remote -> local. Effectively
this silently broke configurations with only a nametrans rule on the
remote folder.

This commit performs both checks, and to all folders, not only to the
folders that are about to be created.

Along the way, store the local and remote names for folders in a
translation table, for convenience and to make the code more readable.

Fixes #408.

Signed-off-by: Abdo Roig-Maranges <abdo.roig@gmail.com>